refactor(#47) :: Notification 패키지 구조 개선 및 NEARBY_ALERT 설정 조정
패키지 리팩토링:
- Notification, NotificationType을 common.entity에서 notification.entity로 이동
- 도메인별 패키지 구조 명확화
- 관련 import 경로 일괄 수정

NEARBY_ALERT 설정 조정:
- 검색 반경: 500m → 1000m (1km로 증가)
- 거리 임계값: 500m → 1000m (같은 장소 판단 기준 완화)
- 더 넓은 범위에서 실종자 감지 가능

영향:
- GPS 업데이트 시 1km 반경 내 실종자 검색
- 이전 알림 위치에서 1km 이내는 같은 장소로 판단
- 알림 빈도는 유지하면서 감지 범위 확대
