+++ src/main/java/baro/baro/domain/notification/dto/NotificationResponse.java
@@ -1,7 +1,7 @@
 package baro.baro.domain.notification.dto;
 
-import baro.baro.domain.common.entity.Notification;
-import baro.baro.domain.common.entity.NotificationType;
+import baro.baro.domain.notification.entity.Notification;
+import baro.baro.domain.notification.entity.NotificationType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
