+++ src/main/java/baro/baro/domain/notification/entity/NotificationType.java
@@ -1,4 +1,4 @@
-package baro.baro.domain.common.entity;
+package baro.baro.domain.notification.entity;
 
 public enum NotificationType {
     INVITE_REQUEST, FOUND_REPORT, NEARBY_ALERT
