+++ src/main/resources/application.properties
@@ -77,6 +77,6 @@ police.api.auth-key=${POLICE_API_AUTH_KEY}
 police.api.row-size=100
 
 # NEARBY_ALERT Configuration
-nearby.alert.radius.meters=${NEARBY_ALERT_RADIUS_METERS:500}
+nearby.alert.radius.meters=${NEARBY_ALERT_RADIUS_METERS:1000}
 nearby.alert.cooldown.hours=${NEARBY_ALERT_COOLDOWN_HOURS:24}
-nearby.alert.distance.threshold.meters=${NEARBY_ALERT_DISTANCE_THRESHOLD_METERS:500}
+nearby.alert.distance.threshold.meters=${NEARBY_ALERT_DISTANCE_THRESHOLD_METERS:1000}
